ItemMenager/pyproject.toml
[tool.poetry]
name = "itemmenager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
twilio = "^9.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
